move gather blocks out of header
